+++ RELEASE_NOTES.md
@@ -2,7 +2,11 @@
 
 ## 1.0.0 (       )
 ### Major Changes
-   * none
+   * Added generic `Procrastinator::Config#log_with`
+      * Removed `Procrastinator::Config#log_inside` 
+      * Removed `Procrastinator::Config#log_at_level`
+      * falsey log level is now the control for whether logging occurs, instead of falsey log directory
+   
 
 ### Minor Changes
    * Started release notes file.